Bugfix for PR #1
PR_#1: Import/export pattern was not implemented correctly under the
components directory
Updated the folder namespaces and implemented the import/export pattern
to resolve the issues.
